refactor: extract interface from LottoPrizeProvider
