+++ README.md
@@ -1,10 +1,9 @@
 # 로또 요구사항
 ## TODO
-
  - 구매한 로또들의 수익률을 구할 수 있다.
- - 등수에 따른 상금을 알 수 있다. 
 
 ## COMPLETED
  - 로또를 만들 수 있다.
  - 구매한 로또와 당첨 번호를 비교해 등수를 구할 수 있다.
  - 로또 번호를 무작위로 뽑을 수 있다.
+ - 등수에 따른 상금을 알 수 있다.
