+++ src/main/java/cc/oakk/lotto/LottoApplication.java
@@ -6,6 +6,8 @@ import java.util.List;
 public class LottoApplication {
     public static void main(String[] args) {
         LottoGenerator generator = new LottoGenerator();
+        LottoPrizeProvider<Long> provider = new SimpleLottoPrizeProvider();
+
         List<Lotto> lottos = new ArrayList<>();
         for (int i = 0; i < 10; i++) {
             lottos.add(generator.generate());
@@ -15,7 +17,7 @@ public class LottoApplication {
         System.out.println();
         lottos.forEach(l -> {
             System.out.println(l);
-            System.out.println(LottoPrizeProvider.getPrizeByRank(l.compare(winningLotto)));
+            System.out.println(provider.getPrizeByRank(l.compare(winningLotto)));
         });
     }
 }
